services/news-alpha-engine/requirements.txt
# ============================================
# NEWS ALPHA ENGINE - Dependencies
# ============================================
# For GPU training (g5.2xlarge with A10G)
# ============================================

# === Core ML ===
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
safetensors>=0.4.0

# === Topic Modeling ===
bertopic>=0.16.0
sentence-transformers>=2.2.0
hdbscan>=0.8.33
umap-learn>=0.5.5

# === GPU Acceleration (cuML) ===
# NOTE: Install via conda for CUDA support:
# conda install -c rapidsai -c conda-forge -c nvidia cuml=24.02
# cuml  # Uncomment if using conda

# === NLP ===
nltk>=3.8.1
spacy>=3.7.0

# === Data Processing ===
pandas>=2.1.0
numpy>=1.26.0
polars>=0.20.0  # Faster than pandas for large datasets
pyarrow>=14.0.0
duckdb>=0.9.0

# === Financial Data ===
yfinance>=0.2.33
polygon-api-client>=1.12.0

# === Training Utilities ===
pytorch-lightning>=2.1.0
wandb>=0.16.0
optuna>=3.5.0  # Hyperparameter tuning
scikit-learn>=1.3.0

# === Explainability ===
shap>=0.44.0
captum>=0.7.0

# === API & Serving ===
fastapi>=0.109.0
uvicorn>=0.25.0
redis>=5.0.0
httpx>=0.26.0
pydantic>=2.5.0

# === Utilities ===
python-dotenv>=1.0.0
structlog>=24.1.0
tqdm>=4.66.0
rich>=13.7.0  # Beautiful terminal output
typer>=0.9.0  # CLI interface
pyyaml>=6.0.1

# === Testing ===
pytest>=7.4.0
pytest-asyncio>=0.23.0

# === Jupyter (for notebooks) ===
jupyter>=1.0.0
ipywidgets>=8.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

